// src/module/entities/model/OrderItem.ts
export class OrderItem {
  identifier: number;
  positionItem: number;
  percentagePis: number;
  percentageCofins: number;
  order: Order;
  itemNumber: number;
  entryDate: string;
  deliveryDate: string;
  product: Product;
  quantity: number;
  stockUnit: {
    unit: string;
    description: string;
    abbreviation: string;
    lastChangeDate: string;
    lastChangeTime: number;
  };
  weight: number;
  packing: {
    code: number;
    description: string;
    abbreviation: string;
    lastChangeDate: string;
    lastChangeTime: number;
  };
  priceList: number;
  price: number;
  standardPrice: number;
  unitCost: number;
  productReference: string;
  productDescription: string;
  purschaseOrder: string;
  itemPurchaseOrder: number;
  externalIdOrder: number;
  combination: number;
  originOfGoods: number;
  movimentedBatch: {
    code: number;
    batchNumber: string;
    supplierBatchNumber: string;
    manufacturingDate: string;
    expirationDate: string;
    availabilityDate: string;
    lastChangeDate: string;
    lastChangeTime: number;
  };
  stockLocation: {
    code: number;
    description: string;
    abbreviation: string;
    lastChangeDate: string;
    lastChangeTime: number;
  };
  firstPercentageDiscount: number;
  secondPercentageDiscount: number;
  thirdPercentageDiscount: number;
  fourthPercentageDiscount: number;
  fifthPercentageDiscount: number;
  percentageIncrease: number;
  icmtCst: number;
  percentageIcms: number;
  percentageIpi: number;
  taxSituationPis: number;
  taxSituationCofins: number;
  taxSituationIpi: number;
  simplesNacionalOperationCode: number;
  grossAmount: number;
  discountAmount: number;
  totalNetAmount: number;
  issPercentageRetained: number;
  perIcmsSubstituteBaseReduction: number;
  freightAmount: number;
  insuranceAmount: number;
  ancillaryExpensesAmount: number;
  individualPackageType: number;
  issPercentageNotRetained: number;
  numberOfLabels: number;
  itemIdObservation: Observation;
  productionIdObservation: Observation;
  lastChangeDate: string;
  lastChangeTime: number;
  cancellationDate: string;
  cancellationTime: number;
  itemPurchaseOrderSiger: number;
}
export class OrderItemFields {
  identifier?: boolean;
  positionItem?: boolean;
  percentagePis?: boolean;
  percentageCofins?: boolean;
  order?: OrderFields;
  itemNumber?: boolean;
  entryDate?: boolean;
  deliveryDate?: boolean;
  product?: ProductFields;
  quantity?: boolean;
  stockUnit?: {
    unit?: boolean;
    description?: boolean;
    abbreviation?: boolean;
    lastChangeDate?: boolean;
    lastChangeTime?: boolean;
  };
  weight?: boolean;
  packing?: {
    code?: boolean;
    description?: boolean;
    abbreviation?: boolean;
    lastChangeDate?: boolean;
    lastChangeTime?: boolean;
  };
  priceList?: boolean;
  price?: boolean;
  standardPrice?: boolean;
  unitCost?: boolean;
  productReference?: boolean;
  productDescription?: boolean;
  purschaseOrder?: boolean;
  itemPurchaseOrder?: boolean;
  externalIdOrder?: boolean;
  combination?: boolean;
  originOfGoods?: boolean;
  movimentedBatch?: {
    code?: boolean;
    batchNumber?: boolean;
    supplierBatchNumber?: boolean;
    manufacturingDate?: boolean;
    expirationDate?: boolean;
    availabilityDate?: boolean;
    lastChangeDate?: boolean;
    lastChangeTime?: boolean;
  };
  stockLocation?: {
    code?: boolean;
    description?: boolean;
    abbreviation?: boolean;
    lastChangeDate?: boolean;
    lastChangeTime?: boolean;
  };
  firstPercentageDiscount?: boolean;
  secondPercentageDiscount?: boolean;
  thirdPercentageDiscount?: boolean;
  fourthPercentageDiscount?: boolean;
  fifthPercentageDiscount?: boolean;
  percentageIncrease?: boolean;
  icmtCst?: boolean;
  percentageIcms?: boolean;
  percentageIpi?: boolean;
  taxSituationPis?: boolean;
  taxSituationCofins?: boolean;
  taxSituationIpi?: boolean;
  simplesNacionalOperationCode?: boolean;
  grossAmount?: boolean;
  discountAmount?: boolean;
  totalNetAmount?: boolean;
  issPercentageRetained?: boolean;
  perIcmsSubstituteBaseReduction?: boolean;
  freightAmount?: boolean;
  insuranceAmount?: boolean;
  ancillaryExpensesAmount?: boolean;
  individualPackageType?: boolean;
  issPercentageNotRetained?: boolean;
  numberOfLabels?: boolean;
  itemIdObservation?: boolean;
  productionIdObservation?: boolean;
  lastChangeDate?: boolean;
  lastChangeTime?: boolean;
}